requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.3.2
python-multipart==0.0.9
pydantic==2.6.0
httpx==0.26.0
